chapters/6-Progress_Bar/pyproject.toml
[tool.poetry]
name = "mandybrot"
version = "0.1.0"
description = "Plot the magical Mandelbrot set."
authors = ["Freddy Wordingham <freddy@digilab.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
numpy = "^1.24.2"
pillow = "^9.4.0"
python = "^3.8"
tdqm = "^0.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
